feat: реализовал best practices и исправил некоторые ошибки
