+++ routes/index.js
@@ -1,0 +1,8 @@
+const router = require('express').Router();
+const users = require('./users');
+const movies = require('./movies');
+
+router.use('/', users);
+router.use('/', movies);
+
+module.exports = router;
