+++ routes/users.js
@@ -1,13 +1,7 @@
 const router = require('express').Router();
-const { celebrate, Joi } = require('celebrate');
+const { celebrate } = require('celebrate');
 const { getCurrentUser, patchUserData } = require('../controllers/users');
-
-const joiObjectUserPatch = {
-  body: Joi.object().keys({
-    name: Joi.string().min(2).max(30).required(),
-    email: Joi.string().email().required(),
-  }),
-};
+const { joiObjectUserPatch } = require('../utils/JoiOptsObjs');
 
 router.get('/users/me', getCurrentUser);
