+++ utils/JoiOptsObjs.js
@@ -1,0 +1,53 @@
+const { Joi } = require('celebrate');
+
+const joiObjectSignup = {
+  body: Joi.object().keys({
+    name: Joi.string().min(2).max(30).required(),
+    email: Joi.string().email().required(),
+    password: Joi.string().required(),
+  }),
+};
+
+const joiObjectSignin = {
+  body: Joi.object().keys({
+    email: Joi.string().email().required(),
+    password: Joi.string().required(),
+  }),
+};
+
+const joiObjectMoviePost = {
+  body: Joi.object().keys({
+    country: Joi.string().required(),
+    director: Joi.string().required(),
+    duration: Joi.number().required(),
+    year: Joi.string().required(),
+    description: Joi.string().required(),
+    image: Joi.string().pattern(/^(?:http(s)?:\/\/)?[\w.-]+(?:\.[\w\.-]+)+[\w\-\._~:/?#[\]@!\$&'\(\)\*\+,;=.]+$/).required(),
+    trailer: Joi.string().pattern(/^(?:http(s)?:\/\/)?[\w.-]+(?:\.[\w\.-]+)+[\w\-\._~:/?#[\]@!\$&'\(\)\*\+,;=.]+$/).required(),
+    nameRU: Joi.string().required(),
+    nameEN: Joi.string().required(),
+    thumbnail: Joi.string().pattern(/^(?:http(s)?:\/\/)?[\w.-]+(?:\.[\w\.-]+)+[\w\-\._~:/?#[\]@!\$&'\(\)\*\+,;=.]+$/).required(),
+    movieId: Joi.string().length(24).hex().required(),
+  }),
+};
+
+const joiObjectMovieDel = {
+  params: Joi.object().keys({
+    movieId: Joi.string().length(24).hex().required(),
+  }),
+};
+
+const joiObjectUserPatch = {
+  body: Joi.object().keys({
+    name: Joi.string().min(2).max(30).required(),
+    email: Joi.string().email().required(),
+  }),
+};
+
+module.exports = {
+  joiObjectSignup,
+  joiObjectSignin,
+  joiObjectMoviePost,
+  joiObjectMovieDel,
+  joiObjectUserPatch,
+};
