+++ utils/centralErrorHandler.js
@@ -1,0 +1,10 @@
+module.exports = (err, req, res, next) => {
+  const { statusCode = 500, message } = err;
+  return res
+    .status(statusCode)
+    .send({
+      message: statusCode === 500
+        ? 'На сервере произошла ошибка. Приносим свои извинения.'
+        : message,
+    });
+};
