+++ utils/rateLimitOpts.js
@@ -1,0 +1,4 @@
+module.exports = {
+  windowMs: 15 * 60 * 1000,
+  max: 100,
+};
